Updated parameter set to be a flat structure

+++ source/util/common_utils.hpp
@@ -17,7 +17,7 @@
 #include <string>
 using namespace std;
 
-typedef double RATETYPE;
+typedef float RATETYPE;
 
 template<class TYPE, int SIZE>
 class CPUGPU_TempArray {
